Ability to Pick Up wind up Key
theres a bug where it picks up both the wind up key AND the Pendellum at the same time. ALSO sometimes you have to press the E key a few times in order for object to be picked up. Also still fixing issue of objects going through floor and walls