Added the ability to move lighter in and out of view by pressing E
Able to pick up JournalBook now and that is required before being able to open journal pages. Did the same thing with the lighter and you are able to move the lighter in and out of view with E without the need of an object in your hand. Likewise when you pick up an item it will put your lighter away.